+++ PCLParaphernalia/ToolFontSamplePCL.cs
@@ -1048,7 +1048,7 @@ namespace PCLParaphernalia
             {
                 posX = (short)(marginX + (_cellWidth * (startIndxMajor + 1)) +
                                           (_cellWidth / 3));
-                posY = (short)(_posYGrid + (_cellHeight * 2 / 3));
+                posY = _posYGrid + (_cellHeight * 2 / 3);
             }
             else
             {
@@ -1126,7 +1126,7 @@ namespace PCLParaphernalia
 
                                 if (codeVal <= _symSetUserMapMax)
                                 {
-                                    mapVal = _symSetUserMap[(int)codeVal];
+                                    mapVal = _symSetUserMap[codeVal];
                                 }
                                 else
                                 {
@@ -1297,7 +1297,7 @@ namespace PCLParaphernalia
                         {
                             if (codeVal <= _symSetUserMapMax)
                             {
-                                mapVal = _symSetUserMap[(int)codeVal];
+                                mapVal = _symSetUserMap[codeVal];
                             }
                             else
                             {
@@ -1377,7 +1377,7 @@ namespace PCLParaphernalia
                         {
                             if (codeVal <= _symSetUserMapMax)
                             {
-                                mapVal = _symSetUserMap[(int)codeVal];
+                                mapVal = _symSetUserMap[codeVal];
                             }
                             else
                             {
