+++ PCLParaphernalia/ToolFontSamplePersist.cs
@@ -378,11 +378,11 @@ namespace PCLParaphernalia
 
                 tmpInt          = (int)subKey.GetValue(_nameHeight,
                                                          _defaultHeightPtsK);
-                height          = (double)(tmpInt / _sizeFactor);
+                height          = tmpInt / _sizeFactor;
 
                 tmpInt          = (int)subKey.GetValue(_namePitch,
                                                          _defaultPitchPtsK);
-                pitch           = (double)(tmpInt / _sizeFactor);
+                pitch           = tmpInt / _sizeFactor;
 
                 indxSymSet      = (int)subKey.GetValue(_nameIndxSymSet,
                                                          _indexZero);
@@ -451,11 +451,11 @@ namespace PCLParaphernalia
 
                 tmpInt       = (int)subKey.GetValue(_nameHeight,
                                                       _defaultHeightPtsK);
-                height       = (double)(tmpInt / _sizeFactor);
+                height       = tmpInt / _sizeFactor;
 
                 tmpInt       = (int)subKey.GetValue(_namePitch,
                                                       _defaultPitchPtsK);
-                pitch        = (double)(tmpInt / _sizeFactor);
+                pitch        = tmpInt / _sizeFactor;
 
                 indxSymSet   = (int)subKey.GetValue(_nameIndxSymSet,
                                                       _indexZero);
@@ -513,11 +513,11 @@ namespace PCLParaphernalia
 
                 tmpInt       = (int)subKey.GetValue(_nameHeight,
                                                       _defaultHeightPtsK);
-                height       = (double)(tmpInt / _sizeFactor);
+                height       = tmpInt / _sizeFactor;
 
                 tmpInt       = (int)subKey.GetValue(_namePitch,
                                                       _defaultPitchPtsK);
-                pitch        = (double)(tmpInt / _sizeFactor);
+                pitch        = tmpInt / _sizeFactor;
 
                 indxSymSet   = (int)subKey.GetValue(_nameIndxSymSet,
                                                       _indexZero);
@@ -654,10 +654,10 @@ namespace PCLParaphernalia
                 weight = (short)tmpInt;
 
                 tmpInt = (int)subKey.GetValue(_nameHeight, _defaultHeightPtsK);
-                height = (double)(tmpInt / _sizeFactor);
+                height = tmpInt / _sizeFactor;
 
                 tmpInt = (int)subKey.GetValue(_namePitch, _defaultPitchPtsK);
-                pitch = (double)(tmpInt / _sizeFactor);
+                pitch = tmpInt / _sizeFactor;
 
                 //------------------------------------------------------------//
 
@@ -707,7 +707,7 @@ namespace PCLParaphernalia
 
                 tmpInt       = (int)subKey.GetValue(_nameHeight,
                                                       _defaultHeightPtsK);
-                height       = (double)(tmpInt / _sizeFactor);
+                height       = tmpInt / _sizeFactor;
 
                 indxSymSet   = (int)subKey.GetValue(_nameIndxSymSet,
                                                       _indexZero);
@@ -765,7 +765,7 @@ namespace PCLParaphernalia
 
                 tmpInt       = (int)subKey.GetValue(_nameHeight,
                                                       _defaultHeightPtsK);
-                height       = (double)(tmpInt / _sizeFactor);
+                height       = tmpInt / _sizeFactor;
 
                 indxSymSet   = (int)subKey.GetValue(_nameIndxSymSet,
                                                       _indexZero);
@@ -820,7 +820,7 @@ namespace PCLParaphernalia
 
                 tmpInt       = (int)subKey.GetValue(_nameHeight,
                                                       _defaultHeightPtsK);
-                height       = (double)(tmpInt / _sizeFactor);
+                height       = tmpInt / _sizeFactor;
 
                 indxSymSet   = (int)subKey.GetValue(_nameIndxSymSet,
                                                       _indexZero);
